+++ CHANGELOG.md
@@ -38,10 +38,14 @@ Other guiding principles:
 
 ## v10.11.20260827 _[unreleased; planned for 2026-08-27]_
 
+
+### Changed
+- **amaru-ledger**: restructured script validations and sped up phase-one validation by up to ~11.5%
+
 ### Fixed
 
 - **workflows**: executable permissions are now correctly preserved in the release workflow.
-- **amaru-ledger**: Correctly calculate an output's minimum lovelace value.
+- **amaru-ledger**: correctly calculate an output's minimum lovelace value.
 - **amaru-ledger**: do not re-encode locally submitted transactions to obtain their size.
 
 ## [v10.11.20260820](https://github.com/pragma-org/amaru/releases/tag/v10.11.20260820)